add swupdate specific systemd-tmpfiles configuration file
The default systemd-tmpfiles configuration for the /tmp directory
(tmp.conf) causes systemd-tmpfiles-clean.service to remove the folders
/tmp/datadst and /tmp/scripts if they are not used for >=10 days.

SWUpdate creates these directories on startup (swupdate_init) and
assumes their existence on use.

For this reason, an SWUpdate specific systemd-tmpfiles configuration
is installed, in order to guarantee that the two directores above
are preserved (only the folders, not their content).

Signed-off-by: Stefano Babic <sbabic@denx.de>
Signed-off-by: Simon Ausserlechner <simon.ausserlechner@siemens.com>